refactor(frontend): group agent studio session actions
